Adding the price modify to Appointment

+++ Views/liste_patient.php
@@ -11,11 +11,7 @@ include("Public/sidebar.php");
   <h2><?php echo ($this->_t)   ?></h2>
 </div>
 
-
-
-
-
-  <div class="col-10 table-responsive">
+<div class="col-10 table-responsive">
     <br>
     <a href="/HMS_PROJECT/ajout_patient" class="btn btn-info float-right"><i class="fas fa-plus-square"></i></a>
     <br>
@@ -38,8 +34,12 @@ foreach($patient as $d):
       <td><?= $d->telephone(); ?></td>
       <td>
         <a href="http://localhost/HMS_PROJECT/index.php?url=detail_patient&id=<?=$d->id();?>" id="link" style="text-decoration:none;"><i class="fas fa-info-circle"></i>  </a>
-        <a href="http://localhost/HMS_PROJECT/index.php?url=delete_patient&id=<?=$d->id();?>" id="link" style="text-decoration:none;"  id="link" style="text-decoration:none;"><i class="fas fa-trash-alt"></i></a>
-        <a href="http://localhost/HMS_PROJECT/index.php?url=modif_patient&id=<?=$d->id();?>" id="link" style="text-decoration:none;"><i class="fas fa-pen-square"></i></a>
+       <?php
+        if(!isset($_COOKIE["doctoremail"])){
+          ?>
+            <a href="http://localhost/HMS_PROJECT/index.php?url=delete_patient&id=<?=$d->id();?>" id="link" style="text-decoration:none;"  id="link" style="text-decoration:none;"><i class="fas fa-trash-alt"></i></a>
+            <a href="http://localhost/HMS_PROJECT/index.php?url=modif_patient&id=<?=$d->id();?>" id="link" style="text-decoration:none;"><i class="fas fa-pen-square"></i></a>
+        <?php } ?>
       </td>
     </tr>
   </tbody>
